backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.15.2
redis[hiredis]==5.3.0
pinecone-client==6.0.2
openai==1.82.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
websockets==15.0.1
tenacity==9.1.2
numpy==2.2.6
tiktoken==0.9.0
pytest==8.4.0
pytest-asyncio==0.26.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
aiohttp==3.12.4
gunicorn==23.0.0
aiosqlite==0.21.0
